feat: Confirm restart on 'env rotate:ENV'

+++ plugins/kubectl/util/save-secrets.sh
@@ -67,9 +67,13 @@ done && \
 
 if [[ ${kubectl_skip_restart:-} != "true" ]]; then
   echo && \
-  echo "--- kubectl: Restarting pods ---" && \
-  echo "TODO rolling update instead of delete?" && \
-  kubectl delete --all pods --namespace="${taito_namespace}"
+  echo "Restart all pods in namespace ${taito_namespace} (Y/n)?" && \
+  read -r confirm && \
+  if [[ "${confirm}" =~ ^[Yy]$ ]]; then
+    echo "--- kubectl: Restarting pods ---" && \
+    echo "TODO rolling update instead of delete?" && \
+    kubectl delete --all pods --namespace="${taito_namespace}"
+  fi
 else
   echo "--- NOTE: Skipping pod restart for refreshing secrets ---"
 fi
